Turn the division switcher into a component
