+++ apps/portal/pages/events/[id]/index.tsx
@@ -1,17 +1,15 @@
 import { NextPage, GetServerSideProps, GetServerSidePropsContext } from 'next';
 import { useTranslations } from 'next-intl';
-import { Box, Container, Stack, Typography, Button, Menu, MenuItem } from '@mui/material';
-import Link from 'next/link';
+import { Container, Typography } from '@mui/material';
 import { PortalEvent, PortalTeam, PortalEventStatus, PortalDivision } from '@lems/types';
 import { fetchAwards, fetchEvent } from '../../../lib/api';
+import DivisionSwitcher from '../../../components/events/division-switcher';
 import EventInfo from '../../../components/events/event-info';
 import EventQuickLinks from '../../../components/events/event-quick-links';
 import TeamList from '../../../components/teams/team-list';
 import EventStatus from '../../../components/events/event-status';
 import { useRealtimeData } from '../../../hooks/use-realtime-data';
 import { getMessages } from '../../../locale/get-messages';
-import React from 'react';
-import { ArrowDropDown } from '@mui/icons-material';
 
 interface Props {
   event: PortalEvent;
@@ -27,77 +25,11 @@ const Page: NextPage<Props> = ({ event, teams, hasAwards, divisions }) => {
     error
   } = useRealtimeData<PortalEventStatus>(`/events/${event.id}/status`);
   const t = useTranslations('pages:events:id:index');
-  const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
-  const open = Boolean(anchorEl);
-
-  const handleClick = (event: React.MouseEvent<HTMLElement>) => {
-    setAnchorEl(event.currentTarget);
-  };
-
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
 
   return (
     <Container maxWidth="md" sx={{ my: 2 }}>
       <Typography variant="h1">{event.name}</Typography>
-      {event.isDivision && (
-        <>
-          <Button
-            variant="text"
-            onClick={handleClick}
-            aria-controls={open ? 'division-menu' : undefined}
-            aria-haspopup="true"
-            aria-expanded={open ? 'true' : undefined}
-            sx={{
-              p: 1,
-              minWidth: 'auto'
-            }}
-          >
-            <Stack direction="row" alignItems="center">
-              <Stack direction="row" spacing={2} alignItems="center">
-                <Box bgcolor={event.color} width={18} height={18} borderRadius={1} />
-                <Typography variant="body1" color="text.secondary">
-                  {event.subtitle}
-                </Typography>
-              </Stack>
-              <ArrowDropDown />
-            </Stack>
-          </Button>
-          <Menu
-            id="division-menu"
-            anchorEl={anchorEl}
-            open={open}
-            onClose={handleClose}
-            MenuListProps={{
-              'aria-labelledby': 'division-button'
-            }}
-            anchorOrigin={{
-              vertical: 'bottom',
-              horizontal: 'left'
-            }}
-            transformOrigin={{
-              vertical: 'top',
-              horizontal: 'left'
-            }}
-          >
-            {divisions?.map(division => (
-              <Link
-                key={division.id}
-                href={`/events/${division.id}`}
-                style={{ textDecoration: 'none', color: 'inherit' }}
-              >
-                <MenuItem onClick={handleClose}>
-                  <Stack direction="row" spacing={2} alignItems="center">
-                    <Box bgcolor={division.color} width={18} height={18} borderRadius={1} />
-                    <Typography variant="body1">{division.name}</Typography>
-                  </Stack>
-                </MenuItem>
-              </Link>
-            ))}
-          </Menu>
-        </>
-      )}
+      {event.isDivision && <DivisionSwitcher event={event} divisions={divisions} />}
       <EventInfo event={event} teamCount={teams.length} />
       {!isLoading && !error && status.isLive && <EventStatus status={status} />}
       <EventQuickLinks event={event} hasAwards={hasAwards} />
